requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Matomo-D==0.0.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
numpy==1.24.2
packaging==23.0
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.29.0
scikit-learn==1.1.1
scipy==1.10.0
sentry-sdk==1.21.0
six==1.16.0
threadpoolctl==3.1.0
typing-extensions==4.4.0
urllib3==1.26.15
Werkzeug==2.2.2
zipp==3.12.1
